pyproject.toml
[project]
name = "orchestration-framework"
version = "0.1.0"
requires-python = ">=3.9"
description = "Multi-agent framework for Snowflake"
authors = [
  { name = "Alejandro Herrera", email = "alejandro.herrera@snowflake.com" },
]
readme = "README.md"
dependencies = [
    "langchain-core>=0.3,<0.3.23",
    "snowflake-snowpark-python[pandas]>=1.22.1,<2.0",
    "asyncio>=3.4.3",
    "aiohttp>=3.10.9,<4.0",
    "pydantic>=2.9.2,<3.0",
    "langchain-text-splitters>=0.3.2",
    "langchain>=0.3.10",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.0.1,<5.0",
    "ruff>=0.6.9,<1.0",
    "pytest>=8.3.3,<9.0",
    "coverage>=7.6.9,<8.0"
]
streamlit = ["streamlit>=1.39.0,<2.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent_gateway"]
